impl. useCreateSecureHttpServer
