+++ example/use_create_secure_http_server.dart
@@ -1,11 +1,12 @@
 import 'package:geoffrey/hooks.dart'
-    show useCreateSecureHttpServer, useRequestHandler, HttpServer, useGet;
+    show useCreateSecureHttpServer, useCertificates, useRequestHandler, useGet;
 
 Future<void> main() async {
   // Creates and returns a new HttpServer and bindSecures it to the given host:port
   // You can do anything with the new server instance
   // HttpServer is the original low level Dart API!
-  HttpServer server = await useCreateSecureHttpServer('localhost', 8080);
+  final sctx = useCertificates(certificate: 'cert.pem', privateKey: 'key.pem');
+  final server = await useCreateSecureHttpServer('localhost', 443, sctx);
 
   // Easy fall back in case hooks are missing some desired functionality
   // If there's something missing please open an issue on GitHub
