+++ lib/hooks.dart
@@ -1,4 +1,4 @@
-export 'dart:io' show HttpRequest, HttpResponse, HttpServer;
+export 'dart:io' show HttpRequest, HttpResponse, HttpServer, SecurityContext;
 
 // Types
 export 'src/globals/context.dart' show Context;
@@ -16,6 +16,7 @@ export 'src/hooks/use_public_dir.dart';
 // server hooks
 export 'src/server_hooks/use_create_http_server.dart';
 export 'src/server_hooks/use_create_secure_http_server.dart';
+export 'src/server_hooks/use_certificates.dart';
 export 'src/server_hooks/use_request_handler.dart';
 export 'src/server_hooks/use_http_server.dart';
