+++ lib/src/helpers/handle_middleware.dart
@@ -1,0 +1,5 @@
+import 'dart:io';
+
+bool handleMiddleware(HttpRequest req, HttpResponse res,
+        Set<bool Function(HttpRequest, HttpResponse)> Middlewares) =>
+    Middlewares.every((middleware) => middleware(req, res));
