+++ lib/src/helpers/map_root_path_to_index_html.dart
@@ -4,6 +4,7 @@ import '../globals/typedefs.dart';
 
 Future<dynamic> mapRootPathToIndexHTml(HttpRequest req) async {
   var index = File('${state[PUBLIC_DIR]}/index.html');
+
   if (await index.exists()) {
     req.response.headers.contentType = ContentType.html;
     return await req.response.addStream(index.openRead());
